it starts

+++ bsock.c
@@ -37,7 +37,7 @@ int bsend(int s, const void *buf, size_t len, int64_t deadline) {
     return b->bsendl(b, &iol, &iol, deadline);
 }
 
-int brecv(int s, void *buf, size_t len, int64_t deadline) {
+ssize_t brecv(int s, void *buf, size_t len, int64_t deadline) {
     struct bsock_vfs *b = hquery(s, bsock_type);
     if(dill_slow(!b)) return -1;
     struct iolist iol = {buf, len, NULL, 0};
@@ -52,7 +52,7 @@ int bsendl(int s, struct iolist *first, struct iolist *last, int64_t deadline) {
     return b->bsendl(b, first, last, deadline);
 }
 
-int brecvl(int s, struct iolist *first, struct iolist *last, int64_t deadline) {
+ssize_t brecvl(int s, struct iolist *first, struct iolist *last, int64_t deadline) {
     struct bsock_vfs *b = hquery(s, bsock_type);
     if(dill_slow(!b)) return -1;
     if(dill_slow((first && !last) || (!first && last) || last->iol_next)) {
